Move oonicli into ui/cli component
