+++ ooni/ui/cli.py
@@ -226,7 +226,7 @@ def setupAnnotations(global_options):
     return annotations
 
 def setupCollector(global_options, collector_client):
-    from backend_client import CollectorClient
+    from ooni.backend_client import CollectorClient
 
     if global_options['collector']:
         collector_client = CollectorClient(global_options['collector'])
@@ -239,7 +239,7 @@ def setupCollector(global_options, collector_client):
 def createDeck(global_options, url=None):
     from ooni.nettest import NetTestLoader
     from ooni.deck import Deck, nettest_to_path
-    from backend_client import CollectorClient
+    from ooni.backend_client import CollectorClient
 
     if url:
         log.msg("Creating deck for: %s" % (url))
